backend/requirements.txt
fastapi
uvicorn
sqlalchemy
python-multipart
python-jose[cryptography]
passlib[bcrypt]
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
openai
python-dotenv
httpx
